@@ -1,19 +1,21 @@
-import "../static/stylesheets/apps.scss";
+import "../../static/stylesheets/apps.scss";
 
 import React from "react";
-import GenericAppLogo from "../static/icons/App.svg";
-import {Action, ImageIcon} from "elv-components-js";
+import GenericAppLogo from "../../static/icons/App.svg";
+import {ImageIcon} from "elv-components-js";
 import UrlJoin from "url-join";
+import {Link} from "react-router-dom";
 
-import FabricBrowserIcon from "../static/images/app_icons/FabricBrowser.png";
-import VideoEditorIcon from "../static/images/app_icons/Video Editor.png";
-import SiteSampleIcon from "../static/images/app_icons/site-sample.svg";
-import StreamSampleIcon from "../static/images/app_icons/stream-sample.svg";
-import StudioIcon from "../static/images/app_icons/Media Ingest.png";
-import ClipSearchIcon from "../static/images/app_icons/clip-search.svg";
-import LiveStreamManagerIcon from "../static/images/app_icons/Livestream Manager.png";
-import CreatorStudioIcon from "../static/images/app_icons/Creator Studio.png";
-import AnalyticsAndReportingIcon from "../static/images/app_icons/Analytics and Reporting.png";
+
+import CreatorStudioIcon from "../../static/images/app_icons/Creator Studio.png";
+import AnalyticsAndReportingIcon from "../../static/images/app_icons/Analytics and Reporting.png";
+import FabricBrowserIcon from "../../static/images/app_icons/FabricBrowser.png";
+import VideoEditorIcon from "../../static/images/app_icons/Video Editor.png";
+import SiteSampleIcon from "../../static/images/app_icons/site-sample.svg";
+import StreamSampleIcon from "../../static/images/app_icons/stream-sample.svg";
+import StudioIcon from "../../static/images/app_icons/Media Ingest.png";
+import ClipSearchIcon from "../../static/images/app_icons/AI Clip Search - beta.png";
+import LiveStreamManagerIcon from "../../static/images/app_icons/Livestream Manager.png";
 
 const icons = {
   "Fabric Browser": FabricBrowserIcon,
@@ -21,7 +23,7 @@ const icons = {
   "Site Sample": SiteSampleIcon,
   "Stream Sample": StreamSampleIcon,
   "Media Ingest": StudioIcon,
-  "Clip Search": ClipSearchIcon,
+  "AI Clip Search": ClipSearchIcon,
   "Livestream Manager": LiveStreamManagerIcon,
   "Creator Studio": CreatorStudioIcon,
   "Eluvio Studio": CreatorStudioIcon,
@@ -32,6 +34,7 @@ const icons = {
 const appNames = [
   "Fabric Browser", "Media Ingest", "Video Editor", "Livestream Manager",
   "Creator Studio", "Eluvio Studio",
+  "AI Clip Search",
   "Analytics and Reporting", "Analytics & Reporting"
 ];
 
@@ -40,12 +43,12 @@ class Apps extends React.PureComponent {
     const logo = icons[Object.keys(icons).find(key => name.includes(key))] || UrlJoin(EluvioConfiguration.apps[name], "Logo.png");
 
     return (
-      <Action key={`app-${name}`} label={`Go to ${name}`} type="link" to={`/apps/${name}`} button={false}>
+      <Link key={`app-${name}`} label={`Go to ${name}`} type="link" to={`/apps/${name}`}>
         <div className="app-selection">
           <ImageIcon icon={logo || GenericAppLogo} alternateIcon={GenericAppLogo} className="app-logo" />
           <h4>{name}</h4>
         </div>
-      </Action>
+      </Link>
     );
   }
 
@@ -57,14 +60,14 @@ class Apps extends React.PureComponent {
       <div className="page-content">
         <div className="apps">
           <div className="apps-box">
-            <h2>Apps</h2>
+            <h2>Content Fabric Application Suite</h2>
             <div className="apps-box__apps">
               { apps.map(name => this.App(name)) }
             </div>
           </div>
 
           <div className="apps-box">
-            <h2>Tools</h2>
+            <h2>Content Fabric Tools</h2>
             <div className="apps-box__apps">
               { tools.map(name => this.App(name)) }
             </div>
