@@ -1,18 +1,23 @@
-import {action, computed, flow, observable} from "mobx";
-import {Buffer} from "buffer";
+import {flow, makeAutoObservable} from "mobx";
+import UrlJoin from "url-join";
 
 class AccountStore {
-  @observable accounts = {};
-  @observable currentAccountAddress;
-  @observable accountsLoaded = false;
+  accounts = {};
+  currentAccountAddress;
+  accountsLoaded = false;
+  tenantAdmins = [];
 
-  @observable loadingAccount;
+  loadingAccount;
 
-  @computed get currentAccount() {
+  get isTenantAdmin() {
+    return this.tenantAdmins.includes(this.currentAccountAddress);
+  }
+
+  get currentAccount() {
     return this.currentAccountAddress ? this.accounts[this.currentAccountAddress] : undefined;
   }
 
-  @computed get sortedAccounts() {
+  get sortedAccounts() {
     return Object.keys(this.accounts)
       .map(address => [address, (this.accounts[address] || {}).name])
       .sort(([addressA, nameA], [addressB, nameB]) => {
@@ -30,6 +35,7 @@ class AccountStore {
   }
 
   constructor(rootStore) {
+    makeAutoObservable(this);
     this.rootStore = rootStore;
 
     this.network = (EluvioConfiguration["config-url"].match(/\.(net\d+)\./) || [])[1] || "";
@@ -43,8 +49,19 @@ class AccountStore {
     });
   }
 
-  @action.bound
   LoadAccounts = flow(function * () {
+    const tenantAdmins = localStorage.getItem(`elv-admins-${this.network}`);
+    if(tenantAdmins) {
+      try {
+        this.tenantAdmins = JSON.parse(tenantAdmins);
+      } catch (error) {
+        // eslint-disable-next-line no-console
+        console.error("Unable to parse tenant admin list:");
+        // eslint-disable-next-line no-console
+        console.error(error);
+      }
+    }
+
     let accounts = localStorage.getItem(`elv-accounts-${this.network}`) || localStorage.getItem("elv-accounts");
     accounts = accounts ? JSON.parse(atob(accounts)) : {};
 
@@ -79,27 +96,27 @@ class AccountStore {
 
     this.accounts = accounts;
 
-    Object.keys(accounts).map(this.AccountBalance);
+    Object.keys(accounts).map(account => this.AccountBalance(account));
 
     this.accountsLoaded = true;
   });
 
-  @action.bound
   AccountBalance = flow(function * (address) {
     const client = this.rootStore.client;
 
     address = client.utils.FormatAddress(address);
 
-    if(!(Object.keys(this.accounts).includes(address))) {
-      return;
-    }
-
-    this.accounts[address].balance = client.utils.ToBigNumber(
+    const balance = client.utils.ToBigNumber(
       yield client.GetBalance({address})
     ).toFixed(3);
+
+    if(Object.keys(this.accounts).includes(address)) {
+      this.accounts[address].balance = balance;
+    }
+
+    return balance;
   });
 
-  @action.bound
   LockAccount({address}) {
     if(!(Object.keys(this.accounts).includes(address))) {
       return;
@@ -108,7 +125,6 @@ class AccountStore {
     this.accounts[address].signer = undefined;
   }
 
-  @action.bound
   UnlockAccount = flow(function * ({address, password}) {
     const client = this.rootStore.client;
     address = client.utils.FormatAddress(address);
@@ -143,18 +159,20 @@ class AccountStore {
     return wallet.GenerateMnemonic();
   }
 
-  @action.bound
   SetTenantContractId = flow(function * ({id}) {
     id = id.trim();
 
     yield this.rootStore.client.userProfileClient.SetTenantContractId({tenantContractId: id});
     this.accounts[this.currentAccountAddress].tenantContractId = yield this.rootStore.client.userProfileClient.TenantContractId();
     yield this.UserMetadata();
+
+    this.SaveAccounts();
   });
 
-  @action.bound
   SetCurrentAccount = flow(function * ({address, signer}) {
     try {
+      this.rootStore.ResetTenancy();
+
       address = this.rootStore.client.utils.FormatAddress(address || signer.address);
 
       this.loadingAccount = address;
@@ -163,6 +181,10 @@ class AccountStore {
 
       if(signer) {
         yield this.rootStore.InitializeClient(signer);
+
+        if(yield this.rootStore.client.userProfileClient.WalletAddress()) {
+          this.accounts[address].tenantContractId = yield this.rootStore.client.userProfileClient.TenantContractId();
+        }
       }
 
       this.accounts[address].signer = signer;
@@ -177,8 +199,8 @@ class AccountStore {
       );
 
       if(signer && this.accounts[address].balance > 0.1) {
-        this.accounts[address].tenantContractId = yield this.rootStore.client.userProfileClient.TenantContractId();
         this.UserMetadata();
+        this.CheckTenantDetails();
       }
     } catch (error) {
       // eslint-disable-next-line no-console
@@ -190,8 +212,18 @@ class AccountStore {
     }
   });
 
-  @action.bound
-  AddAccount = flow(function * ({privateKey, encryptedPrivateKey, mnemonic, password, passwordConfirmation}) {
+  AddAccount = flow(function * ({
+    privateKey,
+    encryptedPrivateKey,
+    mnemonic,
+    password,
+    passwordConfirmation,
+    name,
+    inviteId,
+    adminAddress,
+    faucetToken,
+    tenantContractId
+  }) {
     if(password !== passwordConfirmation) {
       throw Error("Password and confirmation do not match");
     }
@@ -237,15 +269,67 @@ class AccountStore {
       encryptedPrivateKey
     };
 
+    if(inviteId) {
+      yield this.rootStore.tenantStore.ConsumeInvite({
+        tenantContractId,
+        name,
+        address,
+        inviteId,
+        adminAddress,
+        faucetToken
+      });
+    }
+
     yield this.SetCurrentAccount({signer});
+
+    if(name) {
+      yield this.ReplaceUserMetadata({
+        metadataSubtree: UrlJoin("public", "name"),
+        metadata: name
+      });
+    }
+
+    if(tenantContractId) {
+      yield this.SetTenantContractId({
+        id: tenantContractId
+      });
+    }
 
     this.SaveAccounts();
   });
 
+  CheckTenantDetails = flow(function * () {
+    const tenantContractId = yield this.rootStore.client.userProfileClient.TenantContractId();
+    if(tenantContractId) {
+      this.accounts[this.currentAccountAddress].tenantContractId = tenantContractId;
+      const tenantAdminGroupAddress = yield this.rootStore.client.CallContractMethod({
+        contractAddress: this.rootStore.client.utils.HashToAddress(tenantContractId),
+        methodName: "groupsMapping",
+        methodArgs: ["tenant_admin", 0],
+        formatArguments: true,
+      });
+      const accountGroups = yield this.rootStore.client.Collection({collectionType: "accessGroups"});
+
+      const isTenantAdmin = !!accountGroups?.find(address => this.rootStore.client.utils.EqualAddress(tenantAdminGroupAddress, address));
+
+      if(isTenantAdmin) {
+        if(!this.tenantAdmins.includes(this.currentAccountAddress)) {
+          this.tenantAdmins = [...this.tenantAdmins, this.currentAccountAddress];
+        }
+      } else {
+        this.tenantAdmins = this.tenantAdmins.filter(address => !this.rootStore.client.utils.EqualAddress(this.currentAccountAddress, address));
+      }
+
+      localStorage.setItem(`elv-admins-${this.network}`, JSON.stringify(this.tenantAdmins));
+    }
+  });
 
   /* Profile */
 
-  @action.bound
+  async ProfileImage({address}) {
+    return await this.rootStore.client.userProfileClient.UserProfileImage({address});
+  }
+
   UserMetadata = flow(function * () {
     if(!this.currentAccountAddress) { return; }
 
@@ -262,15 +346,7 @@ class AccountStore {
 
     try {
       if(this.accounts[address].metadata.public.profile_image) {
-        this.accounts[address].imageUrl = (yield this.rootStore.client.userProfileClient.UserProfileImage({address}));
-        this.accounts[address].image =
-          "data:image/png;base64," +
-          Buffer.from(
-            yield client.Request({
-              url: this.ResizeImage(this.accounts[address].imageUrl, 200),
-              format: "arrayBuffer"
-            })
-          ).toString("base64");
+        this.accounts[address].imageUrl = (yield this.ProfileImage({address}));
       }
     } catch (error) {
       // eslint-disable-next-line no-console
@@ -284,7 +360,6 @@ class AccountStore {
     this.SaveAccounts();
   });
 
-  @action.bound
   ReplaceUserProfileImage = flow(function * (image) {
     if(!this.currentAccountAddress) { return; }
 
@@ -299,7 +374,6 @@ class AccountStore {
       + `&cache=${Math.random()}` ;
   });
 
-  @action.bound
   ReplaceUserMetadata = flow(function * ({metadataSubtree, metadata}) {
     if(!this.currentAccountAddress) { return; }
 
@@ -308,7 +382,6 @@ class AccountStore {
     yield this.UserMetadata();
   });
 
-  @action.bound
   DeleteUserMetadata = flow(function * ({metadataSubtree}) {
     if(!this.currentAccountAddress) { return; }
 
@@ -317,7 +390,6 @@ class AccountStore {
     yield this.UserMetadata();
   });
 
-  @action.bound
   RemoveAccount(address) {
     if(!(Object.keys(this.accounts).includes(address))) {
       return;
@@ -340,7 +412,8 @@ class AccountStore {
         name: (account.name || "").toString(),
         imageUrl: account.image,
         address: account.address,
-        encryptedPrivateKey: account.encryptedPrivateKey
+        encryptedPrivateKey: account.encryptedPrivateKey,
+        tenantContractId: account.tenantContractId
       }
     );
 
